requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.0.5
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
